Build a self-contained AppImage in CI
Add tools/appimage/build-appimage.sh and an AppImage workflow that
compiles the game, ZMusic, and RTGL from the checkout (clang game,
gcc RTGL, ubuntu-24.04), fetches the release rt/ asset pack from the
RT_ASSETS_URL repository variable (zip or tar, SHA-256 checked),
optionally enables native DLSS from DLSS_SDK_URL, rebuilds
Fluid_Visualize.frag.spv with glslc and overlays it onto the release
shader pack so shaders match the compiled renderer, and packs with
appimage-builder. Triggers on linux-port pushes, PRs touching the
build, published releases (uploading the AppImage), and manual
dispatch. See docs/APPIMAGE_RELEASE.md.

Supporting changes required for the build:

- Use std::pow instead of std::powf for the light-level conversion;
  std::powf is absent from libstdc++ before GCC 13 and breaks the
  clang build (both arguments are float, so std::pow is identical).
- Fetch the RTGL submodule from the acolomba fork over HTTPS, since
  the linux-port gitlinks reference fork-only commits and CI cannot
  authenticate SSH.
- CI environment fixes for ubuntu-24.04: libvulkan-dev/glslc packages
  and a PEP 668 pip install; resolve the rt/ directory (not rt/data)
  and export absolute asset/DLSS paths; provide the Ubuntu archive key
  to appimage-builder's apt venv; pin packaging==21.3; launch the
  gzdoom ELF directly with the asset dir set by AppRun; and install the
  app icon where the bundler looks for it.

+++ src/common/rendering/rt/rt_main.cpp
@@ -1540,7 +1540,7 @@ private:
         {
             return t;
         }
-        return std::powf( t, cvar::rt_lightlevel_exp );
+        return std::pow( t, cvar::rt_lightlevel_exp );
     }
 
     auto MakeFirstPersonQuadInWorldSpace( std::span< const RgPrimitiveVertex > verts )
